+++ src/libfb/if_stack.c
@@ -1,7 +1,7 @@
 /*                      I F _ S T A C K . C
  * BRL-CAD
  *
- * Copyright (c) 1986-2012 United States Government as represented by
+ * Copyright (c) 1986-2014 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * This library is free software; you can redistribute it and/or
@@ -32,6 +32,8 @@
 #include <stdio.h>
 #include <string.h>
 
+#include "bu/log.h"
+#include "bu/str.h"
 #include "fb.h"
 
 
@@ -140,7 +142,7 @@ stk_clear(FBIO *ifp, unsigned char *pp)
 }
 
 
-HIDDEN int
+HIDDEN ssize_t
 stk_read(FBIO *ifp, int x, int y, unsigned char *pixelp, size_t count)
 {
     register FBIO **ip = SI(ifp)->if_list;
@@ -149,11 +151,11 @@ stk_read(FBIO *ifp, int x, int y, unsigned char *pixelp, size_t count)
 	fb_read((*ip), x, y, pixelp, count);
     }
 
-    return (int)count;
+    return count;
 }
 
 
-HIDDEN int
+HIDDEN ssize_t
 stk_write(FBIO *ifp, int x, int y, const unsigned char *pixelp, size_t count)
 {
     register FBIO **ip = SI(ifp)->if_list;
@@ -163,13 +165,11 @@ stk_write(FBIO *ifp, int x, int y, const unsigned char *pixelp, size_t count)
 	ip++;
     }
 
-    return (int)count;
+    return count;
 }
 
 
 /*
- * S T K _ R E A D R E C T
- *
  * Read only from the first source on the stack.
  */
 HIDDEN int
@@ -186,8 +186,6 @@ stk_readrect(FBIO *ifp, int xmin, int ymin, int width, int height, unsigned char
 
 
 /*
- * S T K _ W R I T E R E C T
- *
  * Write to all destinations on the stack
  */
 HIDDEN int
@@ -205,8 +203,6 @@ stk_writerect(FBIO *ifp, int xmin, int ymin, int width, int height, const unsign
 
 
 /*
- * S T K _ B W R E A D R E C T
- *
  * Read only from the first source on the stack.
  */
 HIDDEN int
@@ -223,8 +219,6 @@ stk_bwreadrect(FBIO *ifp, int xmin, int ymin, int width, int height, unsigned ch
 
 
 /*
- * S T K _ B W W R I T E R E C T
- *
  * Write to all destinations on the stack
  */
 HIDDEN int
